fix: conflict

+++ features/content/feed/src/main/java/com/tokopedia/feedplus/view/analytics/FeedEnhancedTracking.kt
@@ -25,7 +25,7 @@ object FeedEnhancedTracking {
 
     fun getClickTracking(
         listPromotion: List<Promotion>,
-        userId: Int
+        userId: Long
     ): Map<String, Any> {
         return DataLayer.mapOf(
             EVENT, Event.PROMO_CLICK,
